Shuffle types around - To ensure we aren't using types in the wrong layers - Moved protocol logic into SignalR - Socket.Abstractions is now the root of the universe, Sockets.Common will likely be removed or turned into Sockets.Common.Http. - Move SSE parser to Sockets.Client and SSE writer into Sockets.Http - Moved tests into the appropriate test projects
