+++ src/Microsoft.AspNetCore.SignalR.Common/Internal/Formatters/BufferExtensions.cs
