+++ src/Microsoft.AspNetCore.SignalR.Common/Internal/Formatters/MessageFormatUtils.cs
