+++ src/Microsoft.AspNetCore.SignalR.Common/Internal/Formatters/TextMessageParser.cs
