+++ src/Microsoft.AspNetCore.SignalR.Common/Message.cs
