+++ src/Microsoft.AspNetCore.SignalR.Common/MessageType.cs
