+++ src/Microsoft.AspNetCore.Sockets.Abstractions/ChannelConnection.cs
