+++ src/Microsoft.AspNetCore.Sockets.Abstractions/IChannelConnection.cs
