+++ src/Microsoft.AspNetCore.Sockets.Client/ServerSentEventsMessageParser.cs
