+++ test/Microsoft.AspNetCore.SignalR.Tests/Formatters/BinaryMessageParserTests.cs
