+++ test/Microsoft.AspNetCore.SignalR.Tests/Formatters/MessageTestUtils.cs
