+++ test/Microsoft.AspNetCore.SignalR.Tests/Formatters/ServerSentEventsParserTests.cs
