+++ test/Microsoft.AspNetCore.SignalR.Tests/Formatters/TextMessageFormatterTests.cs
