+++ test/Microsoft.AspNetCore.Sockets.Tests/ServerSentEventsMessageFormatterTests.cs
